# docs/swagger.yaml
openapi: 3.0.3
info:
  title: AuthzSvc - Enterprise Authentication & Authorization Service
  description: |
    **Production-ready authentication and authorization service** built with Clean Architecture and Hexagonal Pattern principles.
    
    ## Features
    - **JWT-based authentication** with access/refresh token rotation
    - **SMS OTP verification** via Twilio with rate limiting
    - **Role-Based Access Control (RBAC)** with Casbin policy engine
    - **Field-level authorization** with dynamic validation rules
    - **Envoy Proxy integration** for external authorization
    - **Session management** with Redis-based distributed storage
    - **95%+ test coverage** with comprehensive security features
    
    ## Authentication Flow
    1. **Register** user with email/phone
    2. **Verify phone** via SMS OTP
    3. **Login** to get JWT tokens
    4. **Use access token** for API calls
    5. **Refresh tokens** when expired
    
    ## Authorization System
    - **Role inheritance**: admin inherits user permissions
    - **Field validation**: `path.id==token.user_id` ensures ownership
    - **Wildcard patterns**: `/api/users/*` for flexible routing
    - **Runtime policy management** via admin endpoints
    
    ## Security
    - **bcrypt password hashing** (configurable cost)
    - **JWT session validation** with Redis
    - **Rate limiting** on sensitive endpoints
    - **Input validation** and SQL injection prevention
    - **HTTPS/TLS enforcement** in production
...
servers:
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: |
        Register a new user with email, phone, and password. 
        
        **Phone verification required** before login is allowed.
        
        **Default role**: `user` (can be overridden by providing role field)
        
        **Rate limiting**: 10 requests per hour per IP
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic_user:
                summary: Basic user registration
                value:
                  email: user@example.com
                  password: SecurePass123!
                  phone: "+1234567890"
              admin_user:
                summary: Admin user registration
                value:
                  email: admin@example.com
                  password: AdminPass123!
                  phone: "+1234567891"
                  role: admin
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "User registered successfully. Please verify your phone number."
                      user_id:
                        type: integer
                        example: 1
              examples:
                success:
                  summary: Successful registration
                  value:
                    data:
                      message: "User registered successfully. Please verify your phone number."
                      user_id: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "User already exists"
                code: 409
                timestamp: "2024-01-15T10:30:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticate user and receive JWT tokens.
        
        **Requirements**:
        - Valid email and password
        - Phone number must be verified
        - Account must be active
        
        **Returns**:
        - Access token (15min TTL)
        - Refresh token (7 days TTL)  
        - User profile information
        
        **Rate limiting**: 5 requests per 5 minutes per email
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              user_login:
                summary: User login
                value:
                  email: user@example.com
                  password: SecurePass123!
              admin_login:
                summary: Admin login
                value:
                  email: admin@example.com
                  password: AdminPass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthResult'
              examples:
                user_success:
                  summary: Successful user login
                  value:
                    data:
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type: "Bearer"
                      expires_in: 900
                      user:
                        id: 1
                        email: "user@example.com"
                        role: "user"
                admin_success:
                  summary: Successful admin login
                  value:
                    data:
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type: "Bearer"
                      expires_in: 900
                      user:
                        id: 2
                        email: "admin@example.com"
                        role: "admin"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Wrong email or password
                  value:
                    error: "Invalid credentials"
                    code: 401
                    timestamp: "2024-01-15T10:30:00Z"
        '403':
          description: Account inactive or phone not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                account_inactive:
                  summary: Account is inactive
                  value:
                    error: "Account is inactive"
                    code: 403
                    timestamp: "2024-01-15T10:30:00Z"
                phone_not_verified:
                  summary: Phone number not verified
                  value:
                    error: "Phone number not verified"
                    code: 403
                    timestamp: "2024-01-15T10:30:00Z"
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **Format**: `Authorization: Bearer <access_token>`
        
        **Token structure**:
        ```json
        {
          "user_id": 123,
          "role": "user",
          "session_id": "sess_abc123",
          "iss": "authzsvc",
          "iat": 1705320600,
          "exp": 1705321500,
          "jti": "unique-jwt-id"
        }
        ```
        
        **Security notes**:
        - Tokens expire after 15 minutes (configurable)
        - Session must exist in Redis for validation
        - Use refresh endpoint before token expires
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - phone
      properties:
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: user@example.com
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        password:
          type: string
          minLength: 6
          description: Password (minimum 6 characters)
          example: SecurePass123!
          pattern: '^.{6,}$'
        phone:
          type: string
          description: Phone number in international format
          example: "+1234567890"
          pattern: '^\+[1-9]\d{1,14}$'
        role:
          type: string
          description: User role (optional, defaults to 'user')
          example: user
          enum: [user, admin, attendant]
          default: user
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          description: User's password
          example: SecurePass123!
...
    AuthResult:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15min TTL)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          description: JWT refresh token (7 days TTL)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          description: Token type
          example: "Bearer"
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User's email
          example: "user@example.com"
        role:
          type: string
          description: User's role
          example: "user"
          enum: [user, admin, attendant]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code or message
          example: "invalid_credentials"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid email or password"
        code:
          type: integer
          description: HTTP status code
          example: 401
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        request_id:
          type: string
          description: Request ID for tracing
          example: "req_abc123"
        details:
          type: string
          description: Additional error details
          example: "Token validation failed: signature invalid"
  responses:
...
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error
              value:
                error: "validation_failed"
                message: "email: required field missing"
                code: 400
                timestamp: "2024-01-15T10:30:00Z"
            invalid_format:
              summary: Invalid format
              value:
                error: "invalid_format"
                message: "phone number must be in international format"
                code: 400
                timestamp: "2024-01-15T10:30:00Z"
...
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              summary: Rate limit exceeded
              value:
                error: "rate_limit_exceeded"
                message: "Too many requests. Please try again later."
                code: 429
                timestamp: "2024-01-15T10:30:00Z"
            otp_limit:
              summary: OTP rate limit
              value:
                error: "otp_rate_limit"
                message: "Too many OTP requests. Please wait before requesting another."
                code: 429
                timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            general_error:
              summary: General server error
              value:
                error: "internal_server_error"
                message: "An internal error occurred. Please try again later."
                code: 500
                timestamp: "2024-01-15T10:30:00Z"
            database_error:
              summary: Database connection error
              value:
                error: "database_error"
                message: "Database connection failed"
                code: 500
                timestamp: "2024-01-15T10:30:00Z"

# ===== EXAMPLES & DOCUMENTATION =====
...